+++ src/main/java/com/homeaway/devtools/jenkins/testing/WholeClasspathPipelineExtensionDetector.java
@@ -28,7 +28,7 @@ import java.util.Set;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import io.github.lukehutch.fastclasspathscanner.FastClasspathScanner;
+import io.github.classgraph.ClassGraph;
 
 /**
  * Search through classes in the entire classpath for Jenkins Pipeline extensions.
@@ -46,7 +46,11 @@ public class WholeClasspathPipelineExtensionDetector extends APipelineExtensionD
 
 		Set<Class<?>> classes = new HashSet<>();
 
-		List<String> classnames = new FastClasspathScanner(_package.orElse("") ).scan().getNamesOfAllStandardClasses();
+		List<String> classnames = new ClassGraph()
+				.enableAnnotationInfo()
+				.enableClassInfo()
+				.acceptPackages(_package.orElse(""))
+				.scan().getAllStandardClasses().getNames();
 
 		HashMap<String, Throwable> failures = new HashMap<>();
 
@@ -77,10 +81,10 @@ public class WholeClasspathPipelineExtensionDetector extends APipelineExtensionD
 				}
 			} else {
 				LOG.warn(
-					"Failed to get some classes of type [{}] in package [{}]. For detailed error messages, set the system property PipelineExtensionDetector.expandFailures=true. Failures: [{}]",
-					_supertype,
-					_package,
-					failures.keySet() );
+						"Failed to get some classes of type [{}] in package [{}]. For detailed error messages, set the system property PipelineExtensionDetector.expandFailures=true. Failures: [{}]",
+						_supertype,
+						_package,
+						failures.keySet() );
 			}
 		}
 
@@ -94,7 +98,11 @@ public class WholeClasspathPipelineExtensionDetector extends APipelineExtensionD
 
 		HashMap<String, Throwable> failures = new HashMap<>();
 
-		List<String> annotated_classnames = new FastClasspathScanner(_package.orElse("") ).scan().getNamesOfClassesWithAnnotation( _annotation );
+		List<String> annotated_classnames = new ClassGraph()
+				.enableAnnotationInfo()
+				.enableClassInfo()
+				.acceptPackages(_package.orElse(""))
+				.scan().getClassesWithAnnotation( _annotation.getName() ).getNames();
 
 		for(String classname: annotated_classnames) {
 
@@ -123,11 +131,11 @@ public class WholeClasspathPipelineExtensionDetector extends APipelineExtensionD
 				}
 			} else {
 				LOG.warn(
-					"Failed to get some classes annotated with [{}] of type [{}] in package [{}]. For detailed error messages, set the system property PipelineExtensionDetector.expandFailures=true. Failures: [{}]",
-					_annotation,
-					_supertype,
-					_package,
-					failures.keySet() );
+						"Failed to get some classes annotated with [{}] of type [{}] in package [{}]. For detailed error messages, set the system property PipelineExtensionDetector.expandFailures=true. Failures: [{}]",
+						_annotation,
+						_supertype,
+						_package,
+						failures.keySet() );
 			}
 		}